+++ CHANGELOG.md
@@ -13,6 +13,12 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 ### Fixed
+
+
+## 0.2.2
+
+### Fixed
+
 - `git::Repo::last_commit_on_path`: Return `None` when Git log for given path is empty.
 - `cache::Cache::required_object`: Return `None` when required object for at least one input is
   not known.